requirements.txt
fastapi
pillow
pydantic
requests
timm
tokenizers
torch
torchvision
tqdm
transformers
uvicorn
einops
